updating metric apis and pre-commit
